Squash some compiler warnings
git-svn-id: svn+ssh://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx@6145 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/wsprd/wsprd.c
@@ -100,7 +100,7 @@ unsigned long readc2file(char *ptr_to_infile, double *idat, double *qdat,
 //***************************************************************************
 unsigned long readwavfile(char *ptr_to_infile, int ntrmin, double *idat, double *qdat )
 {
-    unsigned long i, j, npoints;
+    int i, j, npoints;
     int nfft1, nfft2, nh2, i0;
     double df;
